FacebookRM: Experimental and buggy support for custom smileys
git-svn-id: http://svn.miranda-ng.org/main/trunk@2734 1316c22d-e87f-b044-9b9b-93d7a3e3ba9c

+++ protocols/FacebookRM/src/utils.h
@@ -93,7 +93,7 @@ namespace utils
 		std::string source_get_value(std::string* data, unsigned int argument_count, ...);
 		std::string source_get_value2(std::string* data, const char *term, const char *endings);
 		void explode(std::string str, std::string separator, std::vector<std::string>* results);
-		void append_ordinal(unsigned int value, std::string* data);
+		void append_ordinal(unsigned long value, std::string* data);
 	};
 
 	namespace conversion
